% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{AWR.S3}
\alias{AWR.S3}
\alias{AWR.S3-package}
\title{An R client to Amazon Batch}
\description{
This is a simple wrapper around the most important feature of the related Java SDK, \code{submitJob()}.
}
\references{
\url{http://docs.aws.amazon.com/AWSJavaSDK/latest/javadoc/com/amazonaws/services/Batch/AWSBatchClient.html}
}
