% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWR-S3.R
\docType{package}
\name{AWR.S3}
\alias{AWR.S3}
\alias{AWR.S3-package}
\alias{s3_put_file}
\alias{s3_get_file}
\alias{s3_delete_file}
\alias{s3_list_files}
\alias{s3_client}
\title{An R client to Amazon S3}
\usage{
s3_put_file(bucket, key, local_file, client = s3_client())

s3_get_file(bucket, key, local_file, client = s3_client())

s3_delete_file(bucket, key, client = s3_client())

s3_list_files(bucket, key, client = s3_client())

s3_client()
}
\arguments{
\item{bucket}{the S3 bucket}

\item{key}{the S3 key}

\item{local_file}{the filename}

\item{client}{SDK client}
}
\description{
This is a simple wrapper around the most important features of the related Java SDK, \code{s3_put_file()}, \code{s3_get_file()}, \code{s3_delete_file()}.

deletes an object from an S3 bucket
}
\references{
\url{http://docs.aws.amazon.com/AWSJavaSDK/latest/javadoc/com/amazonaws/services/Batch/AWSBatchClient.html}
}
