% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit-job.R
\name{submit_job}
\alias{submit_job}
\alias{batch_client}
\title{Submit a Batch Job}
\usage{
submit_job(jobname, jobdefinition, command, jobqueue, vcpu, memory, retries,
  ..., client = batch_client())

batch_client()
}
\arguments{
\item{jobname}{name of the job}

\item{jobdefinition}{name of the job definition}

\item{command}{command and list, presplit}

\item{jobqueue}{name of the job queue}

\item{vcpu}{optional, override the number of vCPU required}

\item{memory}{optional, override the GB of memory required}

\item{retries}{optional, override the number of retries}

\item{...}{not used}

\item{client}{a Batch client (submitJob will be called upon it)}
}
\value{
function result
}
\description{
Submit a Batch Job
}
\examples{
\dontrun{
# Using the default job setup and default auth:
 AWR.S3:::submit_job("test", jobdefinition = "first-run-job-definition", jobqueue = "first-run-job-queue", command = list("echo", "hello"))
}
}
